+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 0.8.0
 - Move data results to field `data` in the resulting object to match GraphQL
+- Support the `query` keyword for graphql
 
 # 0.7.0
 - Added support for complex nested queries when using the `EFRelationshipHandler`